% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crows.R
\docType{data}
\name{crows}
\alias{crows}
\title{PDX Crows: Population Density of Crows in Downtown Portland}
\format{A data frame with 228 rows and 5 variables:
\describe{
  \item{block}{Block (observer defined - not city block) (numerical)}
  \item{date}{Date of data collection (date)}
  \item{side}{Cardinal side of block where crows were spotted (string)}
  \item{count}{Number of crows spotted (numerical)}
  \item{lat}{Latitude (numerical)}
  \item{long}{Longitude (numerical)}
}}
\source{
\email{community-safety@reed.edu}
}
\usage{
crows
}
\description{
A dataset of crow counts in different areas of downtown Portland.
}
\keyword{datasets}
